+++ packages/views/src/index.ts
@@ -5,6 +5,7 @@ export { default as createDataIdHandler } from "./dataIdHandler";
 export * from "./fieldHelpers";
 export { default as createMemoizedHandler } from "./memoizedDataHandler";
 export { default as preventDefault } from "./preventDefault";
+export * from "./routeLink";
 export * from "./types";
 export * from "./useBinding";
 export { default as useDisposable } from "./useDisposable";